api_requirements.txt
PyJWT==2.6.0
pymysql==1.0.3
cryptography==39.0.2
python-dotenv==1.0.0
requests==2.28.2
djangorestframework==3.14.0
django-cors-headers==4.1.0 